feat: add fetched files to build, lot of gardening
