+++ .travis.yml
@@ -4,7 +4,7 @@ branches:
     - develop
 language: node_js
 node_js:
-  - "6"
+  - '8.9.4'
 script:
   - bash ./src/scripts/deploy.sh
 sudo: required
