+++ src/utils/yamlHeadmatter.js
@@ -1,0 +1,11 @@
+module.exports = function yamlHeadmatter(fields) {
+  let ret = '---\n';
+
+  Object.keys(fields).map((field) => {
+    ret += `${ field }: ${ fields[field] }\n`;
+  });
+
+  ret = `${ ret }---\n`
+
+  return ret;
+}
